Move file from subsystem to util
